% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.aggregate.R
\name{ds.aggregate}
\alias{ds.aggregate}
\title{executes some code on some R server sessions.}
\arguments{
\item{connection}{a valid connection to some data repositories. The later needs to be a valid DSConnection-class}

\item{Save}{datashield sessions on each DataSHIELD data repository (if feature is supported) with provided ID (must be a character string).}
}
\description{
Some functions listed as aggregate or based on R can be executed in a DataShield server. Calls DSI::datashield.aggregate function.
}
\author{
Patricia Ryser-Welch
}
