% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.build.login.data.resources.R
\name{ds.build.login.data.resources}
\alias{ds.build.login.data.resources}
\title{Resource assignment}
\usage{
ds.build.login.data.resources(
  servers,
  urls,
  users,
  passwords,
  resources,
  options.ssl,
  drivers
)
}
\arguments{
\item{servers}{A vector of characters listing all the names of the data computers}

\item{urls}{A vector of characters listing each data computer HTTP address. The format is 
\code{https://[TCPIP address or host name][:port]}}

\item{users}{A vector of characters listing a valid user name to log on on each server.}

\item{passwords}{A vector of characters listing the password 
for each user to log in to a data computer.}

\item{resources}{A vector listing the resources name.}

\item{options.ssl}{A vector is used to set the options of the connection. 
Set by defaults the SSL values.}

\item{drivers}{A vector is used to set the name of the driver. 
It is set by default to OpalDriver.}
}
\value{
a data frame. If the arguments are not correct. 
Then a data.frame with no rows is created.

The expectations are as follow:\cr
Expectation no 0: the return value is a data frame\cr
Expectation no 1: the number of columns is equal to 7\cr
Expectation no 2: the number of rows is equal to the number of servers \cr
Expectation no 3: the number of rows is equal to 0, if the length of url, user, or table is 
smaller than the length of the server \cr
Expectation no 4: the number of rows is 0, if any of the URLs do not start with http
}
\description{
Assign a resource object of class "\code{ResourceClient}" to an R symbol in the
DataSHIELD R session.
}
\author{
Patricia Ryser-Welch for DataSHIELD development team
}
