% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.error.R
\name{ds.error}
\alias{ds.error}
\title{displays client-side and server-side error?}
\usage{
ds.error(error, client = TRUE)
}
\arguments{
\item{error}{the error thrown by R}

\item{client}{set to true by default, to indicate the error was thrown by the client-side code. Otherwise, it is set to false, 
to indicate the error was thrown by a server-side code.}
}
\description{
Shows a error message for an error that originated on the server-side calculations or the client-side.
}
