% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.find.variables.R
\name{ds.find.variable}
\alias{ds.find.variable}
\title{indicates whether a variable exists on some DataSHIELD servers}
\arguments{
\item{connection}{a valid connection to some data repositories. The later needs to be a valid DSConnection-class}

\item{variable.name}{name of a variable}
}
\value{
TRUE if the variable exists on every server. FALSE otherwise.
}
\description{
verifies a given varible exists on some DataSHIELD servers, during a session
}
\author{
Patricia Ryser-Welch
}
