% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.login.R
\name{ds.login}
\alias{ds.login}
\title{DataSHIELD servers login}
\usage{
ds.login(
  login.data.frame = NULL,
  assign = TRUE,
  variables = NULL,
  symbol = "D"
)
}
\arguments{
\item{login.data.frame}{a data frame table that holds login details. 
For more information see \strong{Details}.}

\item{assign}{a boolean. If TRUE the data is assigned from the data repository
table to R after login into the server(s). Default TRUE.}

\item{variables}{a vector specifying the variables to assign.
If \code{assign} is set to FALSE this argument is ignored
otherwise the specified variables are assigned to R. 
If no variables are specified (default) the whole
data repository's table is assigned.}

\item{symbol}{a character specifying the name of the data frame to which the data repository's
table will be assigned after login into the server(s). Default \code{"D"}.}
}
\value{
\code{ds.login} returns object(s) of class \code{DSConnection} or NULL
if some parameters are incorrect.
}
\description{
This function uses the login details to log on to 
the specified DataSHIELD server(s).  
The login details can be validated
using the \code{ds.build.login.data.frame} function.
}
\details{
In \code{login.data.frame} table five elements are required to log in to the servers 
where the data to analyse is stored. 
The expected column names are:\cr
\itemize{
  \item \code{driver}: the \code{\link{DSDriver-class}} name, default is \code{OpalDriver} 
  \item \code{server}: the server name
  \item \code{url}: the server URL
  \item \code{user}: the user name or the certificate file path
  \item \code{password}: the user password or the private key file path
  \item \code{table}: the fully qualified name of the table in the data repository
  \item \code{options}: the SSL options
  \item \code{identifiers}: optional column for identifiers mapping 
  (if supported by data repository)
}
 
 To built the mentioned table go to \code{\link{ds.build.login.data.frame}} example. 
 The input table \code{\link{logindata}} for details of the login elements.
}
\author{
Patricia Ryser-Welch for DataSHIELD development team
}
