% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.logout.R
\name{ds.logout}
\alias{ds.logout}
\title{logout from some DataSHIELD servers}
\usage{
ds.logout(connection, save = NULL)
}
\arguments{
\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}

\item{Save}{DataSHIELD sessions on each server data repository (if feature is supported)
with provided ID (must be a character string).}
}
\description{
Clear the DataSHIELD R sessions and logout from DataSHIELD data repositories.
}
\author{
Patricia Ryser-Welch for DataSHIELD development team
}
