% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{ds.share.param}
\alias{ds.share.param}
\title{sharing parameter between}
\usage{
ds.share.param(param.names = NULL, tolerance = 15, datasources = NULL)
}
\arguments{
\item{datasources}{a list of \code{\link{DSConnection-class}} objects obtained after login.}
}
\description{
TODO
}
\author{
Patricia Ryser-Welch for DataSHIELD development
}
