% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.warning.R
\name{ds.warning}
\alias{ds.warning}
\title{Displays DataSHIELD and  R warnings}
\usage{
ds.warning(client.function.name, warning)
}
\arguments{
\item{client.function.name}{a character argument passing the name of the function that has 
thrown the warning}

\item{warning}{The warning thrown by R or a function}
}
\value{
\code{ds.warning} returns a warning either
}
\description{
Shows server-side and/or client-side warnings
}
\seealso{
\code{ds.error}
}
