% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.share.param.R
\name{ds.share.param}
\alias{ds.share.param}
\title{sharing parameter between}
\arguments{
\item{connection}{a valid connection to some data repositories. The later needs to be a valid DSConnection-class}

\item{Save}{datashield sessions on each DataSHIELD data repository (if feature is supported) with provided ID (must be a character string).}
}
\description{
TODO
}
\author{
Patricia Ryser-Welch
}
